requirements.txt
numpy>=1.22,<2.0
pydantic>=2.8.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
prometheus-client>=0.20.0
PyYAML>=6.0
openai-whisper
spacy>=3.7.0
https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl
matplotlib
librosa>=0.10.0,<0.11.0
web3 
py-solc-x 
python-dotenv
gunicorn